rename subpackage material -> materials
