+++ visvis2/renderers/_base.py
@@ -1,6 +1,6 @@
 # todo: all do plural or singular here!
 from visvis2.objects import WorldObject
-from visvis2.material import Material
+from visvis2.materials import Material
 
 
 class Renderer:
