+++ visvis2/renderers/wgpu/linerender.py
@@ -5,7 +5,7 @@ from python_shader import vec4, Array
 
 from . import register_wgpu_render_function, stdinfo_uniform_type
 from ...objects import Line
-from ...material import LineStripMaterial
+from ...materials import LineStripMaterial
 from ..._wrappers import BufferWrapper
 
 # todo: use a compute shader or instancing?
