+++ visvis2/renderers/wgpu/meshrender.py
@@ -3,8 +3,8 @@ import python_shader
 from python_shader import vec4
 
 from . import register_wgpu_render_function, stdinfo_uniform_type
-from ...material._base import Material
 from ...objects import Mesh
+from ...materials import Material
 from ..._wrappers import BufferWrapper
